+++ src/index.ts
@@ -1,6 +1,3 @@
-export * from "./Firebase";
-export * from "./get-config";
-export * from "./useFirebaseContext";
-export * from "./app-check-debug-token";
-export * from "./constants";
-export * from "./types";
+export * from "./FirebaseProvider";
+export * from "./hooks";
+export * from "./util";
