+++ src/util/assignAppCheckDebugToken.ts
