+++ src/util/index.ts
@@ -1,0 +1,5 @@
+export * from "./assignAppCheckDebugToken";
+export * from "./constants";
+export * from "./getConfig";
+export * from "./initializeFirebase";
+export * from "./types";
